refactor wave memory projection
